restructure some components
